+++ index.html
@@ -164,7 +164,7 @@
       </div>
     </section>
   </template>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
