+++ js/create-picture
@@ -1,0 +1,25 @@
+import {getRandomNumber} from './util.js';
+
+const PICTURES_NUMBER = 25;
+const LikesNumber = {
+  MIN: 15,
+  MAX: 200,
+};
+
+const CommentsNumber = {
+  MIN: 0,
+  MAX: 200,
+};
+
+const createPicture = (index) => ({
+  id: index,
+  url: `photos/${ index }.jpg`,
+  description: 'Эта фотография сделана одним из наших авторов',
+  likes: getRandomNumber(LikesNumber.MIN, LikesNumber.MAX),
+  comments: getRandomNumber(CommentsNumber.MIN, CommentsNumber.MAX),
+});
+
+const userPictures = () => Array.from({length: PICTURES_NUMBER}, (_, pictureIndex) => createPicture(pictureIndex + 1));
+
+userPictures();
+export {createPicture};
